refactor ~ replace 'semver' module with `versionCompare()` function
# [why]

The 'semver' module, as used, can be completely removed and replaced with a simple
function, `versionCompare()`, which operates using the built-in `localeCompare()`.

Replacing the 'semver' dependency also increases usability for Deno scripting by avoiding
a prompt for 'env' access permission to 'NODE_DEBUG' (ie, supports *no-panic*/*no-prompt*
behavior for Deno).